+++ src/Common/cellconfig.hpp
@@ -4,53 +4,9 @@
 #include "Model/common_datastructures.hpp"
 #include <cstddef>
 #include <string>
+#include "precision.hpp"
 
 namespace seissol {
-// enum class MaterialType { Static, Acoustic, Elastic, ViscoElastic, PoroElastic, Anisotropic };
-
-enum class PrecisionType {
-  // TODO: add half precision(s) here?
-  Float,
-  Double
-};
-
-template <PrecisionType precision>
-struct PrecisionToType {};
-template <>
-struct PrecisionToType<PrecisionType::Float> {
-  using Type = float;
-};
-template <>
-struct PrecisionToType<PrecisionType::Double> {
-  using Type = double;
-};
-template <typename RealType>
-struct PrecisionFromType {};
-template <>
-struct PrecisionFromType<float> {
-  static constexpr PrecisionType Precision = PrecisionType::Float;
-  static inline const std::string Text = "float";
-};
-template <>
-struct PrecisionFromType<double> {
-  static constexpr PrecisionType Precision = PrecisionType::Double;
-  static inline const std::string Text = "double";
-};
-
-struct CellConfigT {
-  const seissol::model::MaterialType material;
-  const std::size_t mechanisms;
-  const PrecisionType precision;
-  const int convergenceOrder;
-  const bool plasticity;
-
-  bool operator==(const CellConfigT& other) {
-    return material == other.material && mechanisms == other.mechanisms &&
-           precision == other.precision && convergenceOrder == other.convergenceOrder &&
-           plasticity == other.plasticity;
-  }
-};
-
 // TODO(David): once C++20 hits, make CellConfig implement a concept
 // dummy template type
 template <typename MaterialTP, typename RealTP, int ConvergenceOrderP, bool PlasticityP>
